aider: Set the default value of `searchOption` to 'all' if not specified in `listSingStatTables` function.
# Aider chat conversation:

USER: By default, if nothing is specified, searchOption should by default be 'all'

ASSISTANT: Here is the diff to update the `listSingStatTables` function in `singstat/toc.go` to set the default value of `searchOption` to `'all'` if it is not specified:

```diff
--- singstat/toc.go
+++ singstat/toc.go
@@ ... @@
 func listSingStatTables(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	// Retrieve the keyword and searchOption from the query qualifiers
 	quals := d.EqualsQuals
 	keyword := quals["keyword"].GetStringValue()
-	searchOption := quals["searchOption"].GetStringValue()
+	searchOption := "all"
+	if quals["searchOption"] != nil {
+		searchOption = quals["searchOption"].GetStringValue()
+	}

 	// Build the URL for the API request
 	url := "https://tablebuilder.singstat.gov.sg/api/table/resourceid?keyword=" + keyword + "&searchOption=" + searchOption
@@ ... @@
```

This change checks if `searchOption` is provided in the query qualifiers. If not, it defaults to `'all'`. If it is provided, it uses the provided value.